Use uint32 for self destruction setting

+++ src/main/sensors/kaboom.h
@@ -25,7 +25,7 @@ typedef struct kaboomConfig_s {
     uint8_t sensitivity;
     uint8_t more_sensitivity;
     uint16_t activation_time_secs;
-    uint16_t self_destruction_time_secs;
+    uint32_t self_destruction_time_secs;
 } kaboomConfig_t;
 
 PG_DECLARE(kaboomConfig_t, kaboomConfig);
